refactor(web): break dockview state↔component cycle, extract autopick
Two PR-feedback items on #1075:

- Move `enforcePinnedTargets` + `EnforcePinnedTargetsCtx` + the
  `sashDragging` flag from `components/task/dockview-layout-setup.ts`
  into a new `lib/state/dockview-pinned-enforce.ts`. The store and the
  sash-drag handler both now import from this single state-layer
  module; the previous arrangement had `dockview-store` (state) import
  `enforcePinnedTargets` from `dockview-layout-setup` (component),
  which still imported `useDockviewStore` from the state layer - a
  cycle that worked at runtime but crossed the project's
  state-vs-component layer boundary. Flagged by cubic-dev-ai (P2) and
  claude-bot (2x).

- Move `decideAgentProfileAutopick` + `useAgentProfileAutopickEffect`
  + `useWorkflowAgentProfileEffect` from
  `components/task-create-dialog-effects.ts` (was 710 lines, over the
  600 lint limit) into a sibling `task-create-dialog-autopick.ts`.
  Effects file is now 529 lines; the test file imports the moved
  symbols via re-exports kept on the effects module so existing tests
  pass unchanged. Flagged by claude-bot (2x).
